+++ index.ts
@@ -1,32 +1,8 @@
-import express, { Application, Request, Response } from 'express'
+import app from "./app"
 import dotenv from "dotenv"
-import cors from 'cors'
-
 dotenv.config()
-
-
-const app: Application = express();
 const port = process.env.PORT || 4001;
 
-app.use(cors());
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
-app.get("/", (req: Request, res: Response) => {
-    res.status(200).json({
-        message: "Welcome to Politico API"
-    });
-});
-
-app.get("/api/v1", (req:Request, res:Response) => {
-    res.status(200).json({
-        message: "Welcome to Politico API v1. Documentation coming soon."
-    })
-})
-
-
 app.listen(port, () => {
-    console.log(`Server is listening on port ${port}`);
+    console.log(`Server running on port ${port}`)
 })
-
-export default app
